refactor: utils files into utils folder
