+++ apps/journeys-admin/src/components/Editor/Slider/Settings/Drawer/VideoLibrary/VideoFromCloudflare/AddByFile/utils/addByFileUtils.ts
@@ -1,6 +1,6 @@
 import type { ReadStream } from 'fs'
 
-import { CreateCloudflareVideoUploadByFileMutationVariables } from '../../../../../../../../../__generated__/CreateCloudflareVideoUploadByFileMutation'
+import { CreateCloudflareVideoUploadByFileMutationVariables } from '../../../../../../../../../../__generated__/CreateCloudflareVideoUploadByFileMutation'
 
 export function fileToCloudflareUpload(file: File): {
   variables: CreateCloudflareVideoUploadByFileMutationVariables
